del opp contentfetcher i ruter som er brukt internt og ruter som er brukt av andre team
