+++ src/main/kotlin/no/nav/tms/min/side/proxy/Application.kt
@@ -48,15 +48,15 @@ data class AppConfiguration(
         install(HttpTimeout)
     }
 
-    val contentFecther = ContentFetcher(
+    private val proxyHttpClient = ProxyHttpClient(
+        httpClient = httpClient,
         tokendingsService = TokendingsServiceBuilder.buildTokendingsService(),
-        azureService = AzureServiceBuilder.buildAzureService(),
-        aapClientId = aapClientId,
-        aapBaseUrl = aapBaseUrl,
+        azureService = AzureServiceBuilder.buildAzureService()
+    )
+
+    val contentFecther = ContentFetcher(
         eventAggregatorClientId = eventAggregatorClientId,
         eventAggregatorBaseUrl = eventAggregatorBaseUrl,
-        meldekortClientId = meldekortClientId,
-        meldekortBaseUrl = meldekortBaseUrl,
         utkastClientId = utkastClientId,
         utkastBaseUrl = utkastBaseUrl,
         personaliaClientId = personaliaClientId,
@@ -65,9 +65,17 @@ data class AppConfiguration(
         selectorBaseUrl = selectorBaseUrl,
         varselClientId = varselClientId,
         varselBaseUrl = varselBaseUrl,
-        httpClient = httpClient,
         statistikkBaseApiUrl = statistikkBaseUrl,
         statistikkApiId = statistikkClientId,
+        proxyHttpClient = proxyHttpClient,
+    )
+
+    val externalContentFetcher = ExternalContentFetcher(
+        proxyHttpClient = proxyHttpClient,
+        aapClientId = aapClientId,
+        aapBaseUrl = aapBaseUrl,
+        meldekortClientId = meldekortClientId,
+        meldekortBaseUrl = meldekortBaseUrl,
         sykDialogmoteBaseUrl = sykDialogmoteBaseUrl,
         sykDialogmoteClientId = sykDialogmoteClientId,
     )
@@ -86,7 +94,8 @@ fun ApplicationEngineEnvironmentBuilder.envConfig(appConfig: AppConfiguration) {
         proxyApi(
             corsAllowedOrigins = appConfig.corsAllowedOrigins,
             corsAllowedSchemes = appConfig.corsAllowedSchemes,
-            contentFetcher = appConfig.contentFecther
+            contentFetcher = appConfig.contentFecther,
+            externalContentFetcher = appConfig.externalContentFetcher
         )
     }
     connector {
