+++ src/test/kotlin/no/nav/tms/min/side/proxy/ApiTest.kt
@@ -33,11 +33,14 @@ class ApiTest {
         )
 
     @ParameterizedTest
-    @ValueSource(strings = ["aap", "utkast", "personalia", "meldekort", "selector", "varsel","syk/dialogmote"])
+    @ValueSource(strings = ["aap", "utkast", "personalia", "meldekort", "selector", "varsel", "syk/dialogmote"])
     fun `proxy get api`(tjenestePath: String) = testApplication {
         val applicationhttpClient = testApplicationHttpClient()
+        val proxyHttpClient = ProxyHttpClient(applicationhttpClient, tokendigsMock, azureMock)
+
         mockApi(
-            contentFetcher = contentFecther(applicationhttpClient)
+            contentFetcher = contentFecther(proxyHttpClient),
+            externalContentFetcher = externalContentFetcher(proxyHttpClient)
         )
 
         externalServices {
@@ -73,7 +76,9 @@ class ApiTest {
     @ValueSource(strings = ["eventaggregator"])
     fun `proxy post`(tjenestePath: String) = testApplication {
         val applicationhttpClient = testApplicationHttpClient()
-        mockApi(contentFetcher = contentFecther(applicationhttpClient))
+        val proxyHttpClient = ProxyHttpClient(applicationhttpClient, tokendigsMock, azureMock)
+
+        mockApi(contentFetcher = contentFecther(proxyHttpClient), externalContentFetcher = externalContentFetcher(proxyHttpClient))
 
         externalServices {
             hosts(baseurl[tjenestePath]!!) {
@@ -109,7 +114,11 @@ class ApiTest {
     fun `post statistikk`() = testApplication {
         val applicationhttpClient = testApplicationHttpClient()
         var callCount = 0
-        mockApi(contentFetcher = contentFecther(applicationhttpClient))
+        val proxyHttpClient = ProxyHttpClient(applicationhttpClient, tokendigsMock, azureMock)
+        mockApi(
+            contentFetcher = contentFecther(proxyHttpClient),
+            externalContentFetcher = externalContentFetcher(proxyHttpClient)
+        )
 
         externalServices {
             hosts("http://statistikk.test") {
@@ -132,7 +141,11 @@ class ApiTest {
     @Test
     fun healtApiTest() = testApplication {
         val applicationhttpClient = testApplicationHttpClient()
-        mockApi(contentFetcher = contentFecther(applicationhttpClient))
+        val proxyHttpClient = ProxyHttpClient(applicationhttpClient, tokendigsMock, azureMock)
+        mockApi(
+            contentFetcher = contentFecther(proxyHttpClient),
+            externalContentFetcher = externalContentFetcher(proxyHttpClient)
+        )
 
         client.get("/internal/isAlive").status shouldBe HttpStatusCode.OK
         client.get("/internal/isReady").status shouldBe HttpStatusCode.OK
@@ -141,9 +154,10 @@ class ApiTest {
 
     @Test
     fun authPing() = testApplication {
-        mockApi(contentFetcher = mockk())
+        mockApi(contentFetcher = mockk(), externalContentFetcher = mockk())
         client.get("/authPing").status shouldBe HttpStatusCode.OK
     }
+
     private fun checkJson(receiveText: String) {
         if (receiveText == "") throw AssertionError("Post kall har ikke sendt med body")
         try {
@@ -153,26 +167,34 @@ class ApiTest {
         }
     }
 
-    private fun contentFecther(httpClient: HttpClient): ContentFetcher = ContentFetcher(
+    private fun proxyClient(httpClient: HttpClient) = ProxyHttpClient(
+        httpClient = httpClient,
         tokendingsService = tokendigsMock,
-        azureService = azureMock,
-        aapBaseUrl = baseurl["aap"]!!,
-        aapClientId = "aapclient",
+        azureService = azureMock
+    )
+
+    private fun contentFecther(proxyHttpClient: ProxyHttpClient): ContentFetcher = ContentFetcher(
         eventAggregatorClientId = "eventaggregatorclient",
         eventAggregatorBaseUrl = baseurl["eventaggregator"]!!,
         utkastClientId = "utkastclient",
         utkastBaseUrl = baseurl["utkast"]!!,
         personaliaClientId = "personalia",
         personaliaBaseUrl = baseurl["personalia"]!!,
-        meldekortClientId = "meldekort",
-        meldekortBaseUrl = baseurl["meldekort"]!!,
         selectorClientId = "selector",
         selectorBaseUrl = baseurl["selector"]!!,
         varselClientId = "varsel",
         varselBaseUrl = baseurl["varsel"]!!,
-        httpClient = httpClient,
         statistikkApiId = "statistikk",
         statistikkBaseApiUrl = "http://statistikk.test",
+        proxyHttpClient = proxyHttpClient,
+    )
+
+    private fun externalContentFetcher(proxyHttpClient: ProxyHttpClient) = ExternalContentFetcher(
+        proxyHttpClient = proxyHttpClient,
+        aapBaseUrl = baseurl["aap"]!!,
+        aapClientId = "aapclient",
+        meldekortClientId = "meldekort",
+        meldekortBaseUrl = baseurl["meldekort"]!!,
         sykDialogmoteBaseUrl = baseurl["syk/dialogmote"]!!,
         sykDialogmoteClientId = "sykdialogmote",
     )
