+++ src/test/kotlin/no/nav/tms/min/side/proxy/testApplication.kt
@@ -26,12 +26,14 @@ internal fun ApplicationTestBuilder.mockApi(
     corsAllowedOrigins: String = "*.nav.no",
     corsAllowedSchemes: String = "https",
     contentFetcher: ContentFetcher,
+    externalContentFetcher : ExternalContentFetcher,
     securityLevel: SecurityLevel = SecurityLevel.LEVEL_4
 ) = application {
     proxyApi(
         corsAllowedOrigins = corsAllowedOrigins,
         corsAllowedSchemes = corsAllowedSchemes,
         contentFetcher = contentFetcher,
+        externalContentFetcher = externalContentFetcher,
         idportenAuthInstaller = {
             installIdPortenAuthMock {
                 alwaysAuthenticated = true
